fix: resolve PR comments
